refactor(ability): Change cool-down class name and show seconds left
