+++ src/main/resources/assets/narutomod/lang/en_us.json
@@ -48,7 +48,7 @@
   "itemGroup.narutomod_weapons": "Ninja Weapons",
   "itemGroup.narutomod_armor": "Ninja Armor",
   "itemGroup.narutomod_headbands": "Ninja Headbands",
-  "jutsu.fail.cooldown": "%s is on cooldown",
+  "jutsu.fail.cooldown": "%1$s is on cooldown %2$ss",
   "jutsu.fail.notonground": "You must be on the ground to use %s",
   "jutsu.fail.notenoughstamina": "You do not have enough stamina to use %s",
   "jutsu.fail.notenoughchakra": "You do not have enough chakra to use %s",
